+++ demisto_sdk/commands/common/errors.py
@@ -66,8 +66,8 @@ ERROR_CODE = {
     "integration_non_existent_classifier": "IN132",
     "integration_non_existent_mapper": "IN133",
     "multiple_default_arg": "IN134",
-    "missing_output_context": "IN136",
     "invalid_integration_parameters_display_name": "IN135",
+    "missing_output_context": "IN136",
     "invalid_v2_script_name": "SC100",
     "invalid_deprecated_script": "SC101",
     "invalid_command_name_in_script": "SC102",
@@ -454,6 +454,18 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
+    def readme_missing_output_context(command, context_paths):
+        return f'The Following context paths for command {command} are found in YML file ' \
+               f'but are missing from the README file: {context_paths}'
+
+    @staticmethod
+    @error_code_decorator
+    def missing_output_context(command, context_paths):
+        return f'The Following context paths for command {command} are found in the README file ' \
+               f'but are missing from the YML file: {context_paths}'
+
+    @staticmethod
+    @error_code_decorator
     def integration_non_existent_classifier(integration_classifier):
         return f"The integration has a classifier {integration_classifier} which does not exist."
 
@@ -759,7 +771,7 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def missing_release_notes_entry(file_type, pack_name, entity_name):
-        return f"No release note entry was found for the {file_type.lower()} \"{entity_name}\" in the " \
+        return f"No release note entry was found for the {file_type.value.lower()} \"{entity_name}\" in the " \
                f"{pack_name} pack. Please rerun the update-release-notes command without -u to " \
                f"generate an updated template. If you are trying to exclude an item from the release " \
                f"notes, please refer to the documentation found here - " \