+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Added a validation that makes sure outputs in integrations are matching the README file.
 
 # 1.2.19
 * Fixed an issue where merge id_set was not updated to work with the new entity of Packs.