+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 * Do not collect optional dependencies on indicator types reputation commands.
+* Added a validation that makes sure outputs in integrations are matching the README file.
+* Fixed an issue when files were classified incorrectly when running *update-release-notes*.
 
 # 1.2.19
 * Fixed an issue where merge id_set was not updated to work with the new entity of Packs.